update console log message to notifications

+++ src/api/UserInfoRequests/DeleteUserInfoReques.ts
@@ -1,10 +1,14 @@
 import { supabase } from "../supabase-requests";
+import Swal from "sweetalert2";
 
 export const deleteQuizUserInfoFromDb = async (userId: number) => {
-  const { error } = await supabase
-    .from("QuizUsersInfo")
-    .delete()
-    .eq("user_id", userId);
-  if (error){console.log(error);
-    return;}
+  const { error } = await supabase.from("UserInfo").delete().eq("user_id", userId);
+  if (error) {
+    Swal.fire({
+      icon: "error",
+      title: "Oops...",
+      text: "Failed to delete user",
+    });
+    return;
+  }
 };
